+++ tests/test_util.py
@@ -33,7 +33,7 @@ def test_http_get_error():
 
 
 def test_warn_and_assert():
-    with pytest.raises(AssertionError) as excinfo, pytest.warns(UserWarning) as records:
+    with pytest.raises(AssertionError) as excinfo, pytest.warns(UserWarning) as records:  # noqa: PT030
         warn_and_assert([('path/',)], '{0} is invalid', 'See errors above')
 
     assert len(records) == 1